chore: Move ViewModel Files
