+++ feature-onboarding/join/src/main/java/com/apeun/gidaechi/EmailVerificationScreen.kt
@@ -38,6 +38,7 @@ import com.apeun.gidaechi.designsystem.theme.Gray600
 import com.apeun.gidaechi.designsystem.theme.Red500
 import com.apeun.gidaechi.designsystem.theme.SeugiTheme
 import com.apeun.gidaechi.model.EmailVerificationSideEffect
+import com.apeun.gidaechi.viewModel.EmailVerificationViewModel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.collectLatest
